Improve thread safe Utils methods for incrementing identifiers (#3221)

+++ Libraries/Opc.Ua.Server/Diagnostics/DiagnosticsNodeManager.cs
@@ -55,7 +55,7 @@ namespace Opc.Ua.Server
             SetNamespaces(namespaceUris);
 
             m_namespaceIndex = Server.NamespaceUris.GetIndexOrAppend(namespaceUris[1]);
-            m_lastUsedId = DateTime.UtcNow.Ticks & 0x7FFFFFFF;
+            m_lastUsedId = (uint)DateTime.UtcNow.Ticks & 0x7FFFFFFF;
             m_sessions = [];
             m_subscriptions = [];
             DiagnosticsEnabled = true;
@@ -2132,7 +2132,7 @@ namespace Opc.Ua.Server
         }
 
         private readonly ushort m_namespaceIndex;
-        private long m_lastUsedId;
+        private uint m_lastUsedId;
         private Timer m_diagnosticsScanTimer;
         private int m_diagnosticsMonitoringCount;
         private bool m_doScanBusy;
