+++ Libraries/Opc.Ua.Server/Subscription/Subscription.cs
@@ -113,7 +113,7 @@ namespace Opc.Ua.Server
                 MonitoredItemCount = 0,
                 DisabledMonitoredItemCount = 0,
                 MonitoringQueueOverflowCount = 0,
-                NextSequenceNumber = (uint)m_sequenceNumber
+                NextSequenceNumber = m_sequenceNumber
             };
 
             ServerSystemContext systemContext = m_server.DefaultSystemContext.Copy(session);
@@ -194,7 +194,7 @@ namespace Opc.Ua.Server
                 MonitoredItemCount = 0,
                 DisabledMonitoredItemCount = 0,
                 MonitoringQueueOverflowCount = 0,
-                NextSequenceNumber = (uint)m_sequenceNumber
+                NextSequenceNumber = m_sequenceNumber
             };
 
             ServerSystemContext systemContext = m_server.DefaultSystemContext.Copy();
@@ -781,7 +781,7 @@ namespace Opc.Ua.Server
 
                 message = new NotificationMessage
                 {
-                    SequenceNumber = (uint)m_sequenceNumber,
+                    SequenceNumber = m_sequenceNumber,
                     PublishTime = DateTime.UtcNow
                 };
 
@@ -789,7 +789,7 @@ namespace Opc.Ua.Server
 
                 lock (DiagnosticsWriteLock)
                 {
-                    Diagnostics.NextSequenceNumber = (uint)m_sequenceNumber;
+                    Diagnostics.NextSequenceNumber = m_sequenceNumber;
                 }
 
                 var notification = new StatusChangeNotification { Status = StatusCodes.BadTimeout };
@@ -810,7 +810,7 @@ namespace Opc.Ua.Server
             {
                 message = new NotificationMessage
                 {
-                    SequenceNumber = (uint)m_sequenceNumber,
+                    SequenceNumber = m_sequenceNumber,
                     PublishTime = DateTime.UtcNow
                 };
 
@@ -818,7 +818,7 @@ namespace Opc.Ua.Server
 
                 lock (DiagnosticsWriteLock)
                 {
-                    Diagnostics.NextSequenceNumber = (uint)m_sequenceNumber;
+                    Diagnostics.NextSequenceNumber = m_sequenceNumber;
                 }
 
                 var notification = new StatusChangeNotification
@@ -1010,7 +1010,7 @@ namespace Opc.Ua.Server
                 var message = new NotificationMessage
                 {
                     // use the sequence number for the next message.
-                    SequenceNumber = (uint)m_sequenceNumber,
+                    SequenceNumber = m_sequenceNumber,
                     PublishTime = DateTime.UtcNow
                 };
 
@@ -1101,7 +1101,7 @@ namespace Opc.Ua.Server
 
             var message = new NotificationMessage
             {
-                SequenceNumber = (uint)m_sequenceNumber,
+                SequenceNumber = m_sequenceNumber,
                 PublishTime = DateTime.UtcNow
             };
 
@@ -1109,7 +1109,7 @@ namespace Opc.Ua.Server
 
             lock (DiagnosticsWriteLock)
             {
-                Diagnostics.NextSequenceNumber = (uint)m_sequenceNumber;
+                Diagnostics.NextSequenceNumber = m_sequenceNumber;
             }
 
             // add events.
@@ -2714,7 +2714,7 @@ namespace Opc.Ua.Server
         private bool m_waitingForPublish;
         private readonly List<NotificationMessage> m_sentMessages;
         private int m_lastSentMessage;
-        private long m_sequenceNumber;
+        private uint m_sequenceNumber;
         private readonly uint m_maxMessageCount;
         private readonly Dictionary<uint, LinkedListNode<IMonitoredItem>> m_monitoredItems;
         private readonly LinkedList<IMonitoredItem> m_itemsToCheck;
